Game State handling & new font.

+++ Bullet.h
@@ -4,7 +4,7 @@
 class Bullet {
 private:
 	//Variables
-	int type;
+	unsigned short type;
 	sf::Sprite bubble;
 
 public:
